+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Cap XP gain from XP macro at level 20
+
 ## [2.0.2] - 2024-10-02
 
 ### Fixed